+++ com.unity.visualeffectgraph/Editor/Models/Contexts/Implementations/VFXAbstractParticleOutput.cs
@@ -31,7 +31,8 @@ namespace UnityEditor.VFX
             Simple,
             Flipbook,
             FlipbookBlend,
-            ScaleAndBias
+            ScaleAndBias,
+            FlipbookMotionBlend,
         }
 
         public enum ZWriteMode
@@ -136,7 +137,7 @@ namespace UnityEditor.VFX
 
         protected bool isBlendModeOpaque { get { return blendMode == BlendMode.Opaque || blendMode == BlendMode.Masked; } }
 
-        protected bool usesFlipbook { get { return supportsUV && (uvMode == UVMode.Flipbook || uvMode == UVMode.FlipbookBlend); } }
+        protected bool usesFlipbook { get { return supportsUV && (uvMode == UVMode.Flipbook || uvMode == UVMode.FlipbookBlend || uvMode == UVMode.FlipbookMotionBlend); } }
 
         protected virtual IEnumerable<VFXNamedExpression> CollectGPUExpressions(IEnumerable<VFXNamedExpression> slotExpressions)
         {
@@ -157,13 +158,20 @@ namespace UnityEditor.VFX
 
             if (supportsUV && uvMode != UVMode.Simple)
             {
+                VFXNamedExpression flipBookSizeExp;
                 switch (uvMode)
                 {
                     case UVMode.Flipbook:
                     case UVMode.FlipbookBlend:
-                        var flipBookSizeExp = slotExpressions.First(o => o.name == "flipBookSize");
+                    case UVMode.FlipbookMotionBlend:
+                        flipBookSizeExp = slotExpressions.First(o => o.name == "flipBookSize");
                         yield return flipBookSizeExp;
                         yield return new VFXNamedExpression(VFXValue.Constant(Vector2.one) / flipBookSizeExp.exp, "invFlipBookSize");
+                        if (uvMode == UVMode.FlipbookMotionBlend)
+                        {
+                            yield return slotExpressions.First(o => o.name == "motionVectorMap");
+                            yield return slotExpressions.First(o => o.name == "motionVectorScale");
+                        }
                         break;
                     case UVMode.ScaleAndBias:
                         yield return slotExpressions.First(o => o.name == "uvScale");
@@ -210,7 +218,13 @@ namespace UnityEditor.VFX
                     {
                         case UVMode.Flipbook:
                         case UVMode.FlipbookBlend:
+                        case UVMode.FlipbookMotionBlend:
                             yield return new VFXPropertyWithValue(new VFXProperty(typeof(Vector2), "flipBookSize"), new Vector2(4, 4));
+                            if(uvMode == UVMode.FlipbookMotionBlend)
+                            {
+                                yield return new VFXPropertyWithValue(new VFXProperty(typeof(Texture2D), "motionVectorMap"));
+                                yield return new VFXPropertyWithValue(new VFXProperty(typeof(float), "motionVectorScale"), 1.0f);
+                            }
                             break;
                         case UVMode.ScaleAndBias:
                             yield return new VFXPropertyWithValue(new VFXProperty(typeof(Vector2), "uvScale"), Vector2.one);
@@ -287,6 +301,11 @@ namespace UnityEditor.VFX
                         case UVMode.FlipbookBlend:
                             yield return "USE_FLIPBOOK";
                             yield return "USE_FLIPBOOK_INTERPOLATION";
+                            break;
+                        case UVMode.FlipbookMotionBlend:
+                            yield return "USE_FLIPBOOK";
+                            yield return "USE_FLIPBOOK_INTERPOLATION";
+                            yield return "USE_FLIPBOOK_MOTIONVECTORS";
                             break;
                         case UVMode.ScaleAndBias:
                             yield return "USE_UV_SCALE_BIAS";