+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [6.7.0-preview] - 2019-XX-XX
 ### Added
 - Distortion Outputs (Quad / Mesh)
+- Motion-Vector-Based Flipbook Interpolation
 - Color mapping mode for unlit outputs (Textured/Gradient Mapped)
 - Add Triangle and Octagon primitives for particle outputs